Login component split
