+++ frontend/src/Components/LoginPage/Login.css
@@ -13,6 +13,7 @@
 	background-color: whitesmoke;
 	border-radius: 15px;
 	height: 65vh;
+	min-height: 300px;
 	width: 65vw;
 }
